Starting problem 77

+++ Computing/Combination.cpp
@@ -10,11 +10,11 @@ namespace Computing
             return n * Factorial(n-1);
     }
 
-    void Combine(set<int> & coins, set<int>::iterator it, int val, long long & count)
+    void Combine(set<long long> & coins, set<long long>::iterator it, long long val, long long & count)
     {
         while(it != coins.end())
         {
-            int res = val - *it;
+            long long res = val - *it;
             if(res == 0)
                 ++count;
             else if(res > 0)
